combine system and functions, fix outdated api
